[MIG][12.0] pos_payment_terminal
