+++ pos_payment_terminal/__manifest__.py
@@ -7,15 +7,16 @@
     'version': '12.0.0.1.1',
     'category': 'Point Of Sale',
     'summary': 'Manage Payment Terminal device from POS front end',
-    'author': "Aurélien DUMAINE,Akretion,Odoo Community Association (OCA)",
+    'author': "Aurélien DUMAINE,GRAP,Akretion,"
+              "Odoo Community Association (OCA)",
     'license': 'AGPL-3',
     'depends': ['point_of_sale'],
     'data': [
         'views/pos_config.xml',
         'views/account_journal.xml',
-        'views/pos_payment_terminal_template.xml',
+        'views/assets.xml',
         ],
-    'demo': ['demo/pos_payment_terminal_demo.xml'],
+    'demo': ['demo/pos_config.xml'],
     'qweb': ['static/src/xml/pos_payment_terminal.xml'],
     'installable': True,
 }
