+++ pos_payment_terminal/demo/pos_config.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="point_of_sale.pos_config_main" model="pos.config">
+        <field name="iface_payment_terminal" eval="True"/>
+    </record>
+
+</odoo>
