--- pos_payment_terminal/demo/pos_payment_terminal_demo.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-
-<record id="point_of_sale.pos_config_main" model="pos.config">
-    <field name="iface_payment_terminal" eval="True"/>
-</record>
-
-</odoo>
