+++ pos_payment_terminal/i18n/pos_payment_terminal.pot
@@ -1,17 +1,24 @@
 # Translation of Odoo Server.
 # This file contains the translation of the following modules:
-#	* pos_payment_terminal
+#   * pos_payment_terminal
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 10.0\n"
+"Project-Id-Version: Odoo Server 12.0\n"
 "Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2020-01-18 10:55+0000\n"
+"PO-Revision-Date: 2020-01-18 10:55+0000\n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
 "Plural-Forms: \n"
+
+#. module: pos_payment_terminal
+#: model:ir.model.fields,help:pos_payment_terminal.field_pos_config__iface_payment_terminal
+msgid "A payment terminal is available on the Proxy"
+msgstr ""
 
 #. module: pos_payment_terminal
 #: selection:account.journal,payment_mode:0
@@ -29,8 +36,19 @@ msgid "Journal"
 msgstr ""
 
 #. module: pos_payment_terminal
-#: model:ir.model.fields,field_description:pos_payment_terminal.field_account_journal_payment_mode
+#: model:ir.model.fields,field_description:pos_payment_terminal.field_account_journal__payment_mode
 msgid "Payment Mode"
+msgstr ""
+
+#. module: pos_payment_terminal
+#: model:ir.model.fields,field_description:pos_payment_terminal.field_pos_config__iface_payment_terminal
+#: model_terms:ir.ui.view,arch_db:pos_payment_terminal.view_pos_config_form
+msgid "Payment Terminal"
+msgstr ""
+
+#. module: pos_payment_terminal
+#: model:ir.model,name:pos_payment_terminal.model_pos_config
+msgid "Point of Sale Configuration"
 msgstr ""
 
 #. module: pos_payment_terminal
@@ -39,7 +57,7 @@ msgid "Point of Sale Orders"
 msgstr ""
 
 #. module: pos_payment_terminal
-#: model:ir.model.fields,help:pos_payment_terminal.field_account_journal_payment_mode
+#: model:ir.model.fields,help:pos_payment_terminal.field_account_journal__payment_mode
 msgid "Select the payment mode sent to the payment terminal"
 msgstr ""
 
@@ -49,9 +67,3 @@ msgstr ""
 #, python-format
 msgid "Start transaction"
 msgstr ""
-
-#. module: pos_payment_terminal
-#: model:ir.model,name:pos_payment_terminal.model_pos_config
-msgid "pos.config"
-msgstr ""
-
