+++ pos_payment_terminal/models/account_journal.py
@@ -2,7 +2,6 @@
 # © 2015-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-
 from odoo import models, fields
