+++ pos_payment_terminal/models/pos_order.py
@@ -1,9 +1,10 @@
 # © 2018 ACSONE SA/NV
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 from collections import defaultdict
 import logging
 
-from odoo import models, api
+from odoo import api, models
 
 _logger = logging.getLogger(__name__)
