+++ pos_payment_terminal/readme/CONFIGURE.rst
@@ -1,4 +1,11 @@
+* Go to Point of Sale > Configuration > Point of Sale, and check
+  the field "Payment Terminal"
 
-This module support two payment modes : *card* and *check*. The payment
-mode should be configured on the main Odoo server, in the menu *Point
-of Sale > Configuration > Payment Methods*, under the *Point of Sale* tab.
+.. figure:: ../static/description/pos_config_form.png
+
+* Then, in the same point of sale form, click on "Payment Methods" to
+  set correctly your account journals.
+  This module support two payment modes : *card* and *check*, this
+  option is available in te Point of sale tab.
+
+.. figure:: ../static/description/account_journal_form.png
