+++ pos_payment_terminal/readme/CONTRIBUTORS.rst
@@ -1,2 +1,3 @@
 * Aurelien Dumaine
 * Alexis de Lattre <alexis.delattre@akretion.com>
+* Sylvain LE GAL (https://twitter.com/legalsylvain)
